+++ example/android/app/src/main/kotlin/com/ko2ic/imagedownloader/example/MainActivity.kt
@@ -2,4 +2,4 @@ package com.ko2ic.imagedownloader.example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity : FlutterActivity()
+class MainActivity: FlutterActivity()